✅ Create a new tests project for integration testing
