+++ Mydemenageur.API/Mydemenageur.API.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile>C:\Users\Tom\source\repos\api\Mydemenageur.API\Mydemenageur.API.xml</DocumentationFile>
+    <DocumentationFile>C:\Users\victo\Documents\Projets\API\Mydemenageur\Mydemenageur.API\Mydemenageur.API.xml</DocumentationFile>
     <NoWarn>1701;1702;1591;1587</NoWarn>
   </PropertyGroup>
