+++ Mydemenageur.IntegrationTests/Helpers/TestApplicationFactory.cs
@@ -1,0 +1,34 @@
+﻿using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Mydemenageur.IntegrationTests.Helpers
+{
+    public class TestApplicationFactory<TStartup, TTestStartup> : WebApplicationFactory<TTestStartup> where TTestStartup : class
+    {
+        protected override IHostBuilder CreateHostBuilder()
+        {
+            var host = Host.CreateDefaultBuilder()
+                            .ConfigureWebHost(builder =>
+                            {
+                                builder.UseStartup<TTestStartup>();
+                            })
+                            .ConfigureAppConfiguration((context, conf) =>
+                            {
+                                var projectDir = Directory.GetCurrentDirectory();
+                                var configPath = Path.Combine(projectDir, "appsettings.json");
+
+                                conf.AddJsonFile(configPath);
+                            });
+
+            return host;
+        }
+    }
+}
