+++ Mydemenageur.IntegrationTests/appsettings.json
@@ -1,0 +1,24 @@
+{
+  "MongoSettings": {
+    "UsersCollectionName": "users",
+    "ClientsCollectionName": "clients",
+    "MoversCollectionName": "movers",
+    "SocietiesCollectionName": "socities",
+    "VehiculesCollectionName": "vehicules",
+    "HousingsCollectionName": "housings",
+    "MoveRequestsCollectionName": "moverequests",
+    "ConnectionString": "mongodb://localhost:27017",
+    "DatabaseName": "MydemenageurTestDb"
+  },
+  "MydemenageurSettings": {
+    "ApiSecret": "fAePd2iB8hQfzBqgOzbw88gEiWGbI"
+  },
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  },
+  "AllowedHosts": "*"
+}
