--- Mydemenageur.Test/IntegrationsTest/AuthenticationTestService.cs
@@ -1,56 +1,0 @@
-﻿using Microsoft.IdentityModel.Tokens;
-using MongoDB.Driver;
-using Mydemenageur.API.Entities;
-using Mydemenageur.API.Models.Users;
-using Mydemenageur.API.Services.Interfaces;
-using Mydemenageur.API.Settings.Interfaces;
-using System;
-using System.Collections.Generic;
-using System.IdentityModel.Tokens.Jwt;
-using System.Linq;
-using System.Security.Claims;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Mydemenageur.API.Services
-{
-    public class AuthenticationTestService
-    {
-        private readonly IMongoCollection<User> _users;
-
-        private readonly IMydemenageurSettings _mydemenageurSettings;
-
-        public AuthenticationTestService(IMongoSettings mongoSettings, IMydemenageurSettings mydemenageurSettings)
-        {
-            var mongoClient = new MongoClient(mongoSettings.ConnectionString);
-            var database = mongoClient.GetDatabase(mongoSettings.DatabaseName);
-
-            _users = database.GetCollection<User>(mongoSettings.UsersCollectionName);
-
-            _mydemenageurSettings = mydemenageurSettings;
-        }
-
-        // Generate the token for the user
-        public string TokenForUser(string Id, string Role)
-        {
-            var tokenHandler = new JwtSecurityTokenHandler();
-            var key = Encoding.ASCII.GetBytes(_mydemenageurSettings.ApiSecret);
-            var tokenDescriptor = new SecurityTokenDescriptor
-            {
-                Subject = new ClaimsIdentity(new Claim[]
-                {
-                    new Claim(ClaimTypes.Name, Id),
-                    new Claim(ClaimTypes.Role, Role)
-                }),
-                Expires = DateTime.UtcNow.AddDays(20),
-                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
-            };
-
-            var token = tokenHandler.CreateToken(tokenDescriptor);
-
-            return tokenHandler.WriteToken(token);
-        }
-    }
-
-
-}
