+++ Mydemenageur.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 16.0.31005.135
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mydemenageur.API", "Mydemenageur.API\Mydemenageur.API.csproj", "{B2BFABAD-DB46-4F4B-B3DC-38EC5959BE8E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mydemenageur.Test", "Mydemenageur.Test\Mydemenageur.Test.csproj", "{9E415C7E-7AD4-435B-8B6B-E6C8855FD1B8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mydemenageur.IntegrationTests", "Mydemenageur.IntegrationTests\Mydemenageur.IntegrationTests.csproj", "{B4FD2543-DF4D-4F97-94DF-9EEBE4EE0EE4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,10 @@ Global
 		{B2BFABAD-DB46-4F4B-B3DC-38EC5959BE8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B2BFABAD-DB46-4F4B-B3DC-38EC5959BE8E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B2BFABAD-DB46-4F4B-B3DC-38EC5959BE8E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9E415C7E-7AD4-435B-8B6B-E6C8855FD1B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9E415C7E-7AD4-435B-8B6B-E6C8855FD1B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9E415C7E-7AD4-435B-8B6B-E6C8855FD1B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9E415C7E-7AD4-435B-8B6B-E6C8855FD1B8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B4FD2543-DF4D-4F97-94DF-9EEBE4EE0EE4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B4FD2543-DF4D-4F97-94DF-9EEBE4EE0EE4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B4FD2543-DF4D-4F97-94DF-9EEBE4EE0EE4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B4FD2543-DF4D-4F97-94DF-9EEBE4EE0EE4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
